requirements_dev.txt
# Generic python dependencies
pip==25.3
pre-commit==4.5.1
ruff==0.14.10
ty==0.0.7

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.300
syrupy>=4.6.1

# Component dependencies
hassil>=3.2.0
home-assistant-intents>=2025.9.24
mutagen>=1.47.0

google-adk==1.21.0
